backend/app/core/config.py
import os
from dataclasses import dataclass
from pathlib import Path

from dotenv import load_dotenv

BASE_DIR = Path(__file__).resolve().parents[2]
load_dotenv(BASE_DIR / ".env")

# Temporary debug prints for environment verification.
print("DB HOST:", os.getenv("DB_HOST"))
print("DB USER:", os.getenv("DB_USER"))
print("DB PASSWORD:", os.getenv("DB_PASSWORD"))


@dataclass(frozen=True)
class Settings:
    db_host: str = os.getenv("DB_HOST", "127.0.0.1")
    db_port: int = int(os.getenv("DB_PORT", "3306"))
    db_user: str = os.getenv("DB_USER", "root")
    db_password: str = os.getenv("DB_PASSWORD", "")
    db_name: str = os.getenv("DB_NAME", "upi_system")


settings = Settings()
