backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
click==8.3.2
fastapi==0.135.3
h11==0.16.0
idna==3.11
mysql-connector-python==9.6.0
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
